Carry over security statements into the documentation of queries
