+++ app/src/main/kotlin/de/fraunhofer/aisec/openstack/queries/OpenstackComponents.kt
@@ -1,0 +1,9 @@
+/*
+ * This file is part of the OpenStack Checker
+ */
+package de.fraunhofer.aisec.openstack.queries
+object OpenStackComponents {
+    const val KEYSTONE = "keystone"
+    const val CINDER = "cinder"
+    const val BARBICAN = "barbican"
+}
