+++ app/src/main/kotlin/de/fraunhofer/aisec/openstack/queries/encryption/GoodEncryption.kt
@@ -16,9 +16,6 @@ import de.fraunhofer.aisec.cpg.query.*
  * efficiency"
  */
 fun stateOfTheArtEncryption(tr: TranslationResult): QueryTree<Boolean> {
-    // We currently allow the two ciphers aes-xts-plain64 and aes-cbc-essiv.
-    // This could be extracted to a variable outside this statement.
-    val allowedCiphers = listOf("aes-xts-plain64", "aes-cbc-essiv")
 
     // The predicate must hold for all DiskEncryption concepts.
     return tr.allExtended<DiskEncryption> {
