--- projects/magnum-cve/queries/delete-secrets.query.kts
@@ -1,1 +1,0 @@
-../../common/queries/delete-secrets.query.kts
