--- projects/magnum-cve/queries/no-logging-of-secrets.query.kts
@@ -1,1 +1,0 @@
-../../common/queries/no-logging-of-secrets.query.kts
